backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
pdfplumber>=0.10
python-docx>=1.1
spacy>=3.7
scikit-learn>=1.4
numpy>=1.26
anthropic>=0.20
Pillow>=10.0
